feat: tree folder visualization
This is a feature that shows the hierarchy for the current folder, I've named it "missogates" and used an airplane cause is like going somewhere and using a specific gate to reach the destination.

To use is very simple, just click on the airplane then should appear a hyperlink for every children's folder for the current one. If none then a message should appear telling you there are no gates. Click it again to make it disappear.